+++ src/firewallfabrik/platforms/nftables/__init__.py
@@ -12,4 +12,4 @@
 
 """nftables platform compiler package."""
 
-__compiler_version__ = '1.5.17'
+__compiler_version__ = '1.5.18'